exminate the duplicated codes
