+++ kraken-app/kraken-app-portal/src/pages/NewAPIMapping/components/ResponseItem/index.tsx
@@ -16,6 +16,7 @@ import { useEffect, useState } from "react";
 import { useBoolean } from "usehooks-ts";
 import { SourceInput } from "./SourceInput";
 import { TargetInput } from "./TargetInput";
+import { handleDeleteMappingItems } from "@/pages/NewAPIMapping/helper";
 import styles from "./index.module.scss";
 
 type Props = {
@@ -88,24 +89,7 @@ const ResponseItem = ({ item, index }: Props) => {
   };
   
   const handleDeleteMapping = (key: React.Key) => {
-    const targetItem = listMapping.find((item) => item.key === key);
-    if (!targetItem) {
-      return;
-    }
-    // Filter out the deleted item
-    const filtered = listMapping.filter((item) => item.key !== key);
-    // Check if this was the last item of its group (by name)
-    const remainingGroupItems = filtered.filter(item => item.name === targetItem.name);
-    const updated = [...filtered];
-    if (remainingGroupItems.length === 0) {
-      // Re-add a placeholder entry with empty mapping
-      updated.push({
-        name: targetItem.name,
-        key: nanoid(), // generate a new unique key
-        from: undefined,
-        to: [],
-      });
-    }
+    const updated = handleDeleteMappingItems(key, listMapping, []);
     setListMappingStateResponse(updated);
   };
